토큰 인증 실패 처리 계층을 수정하라
토큰을 인코딩하고 디코딩하는 과정에서 실패에 대한 메세지를 응답하고 있었습니다. 하지만 해당 클래스는 인코딩과 디코딩하는 일에 충실하게 만드는 것이 좋다고 생각합니다. 그래서, 인증을 수행할 때 실패를 인증하도록 수정했습니다
