+++ app/src/main/java/com/codesoom/assignment/errors/AuthenticationBadRequestException.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.errors;
+
+public class AuthenticationBadRequestException extends RuntimeException{
+    public AuthenticationBadRequestException() {
+        super("Authentication bad request");
+    }
+}
